Add provider option to PrismaAdapter (#4009)

+++ packages/adapter-prisma/lib/adapter-prisma.js
@@ -10,6 +10,7 @@ class PrismaAdapter extends BaseKeystoneAdapter {
   constructor() {
     super(...arguments);
     this.name = 'prisma';
+    this.provider = this.config.provider || 'postgresql';
 
     this.getPrismaPath = this.config.getPrismaPath || (() => '.prisma');
     this.getDbSchemaName = this.config.getDbSchemaName || (() => 'public');
@@ -22,7 +23,7 @@ class PrismaAdapter extends BaseKeystoneAdapter {
     const { PrismaClient } = require(this.clientPath);
     this.prisma = new PrismaClient({
       log: this.enableLogging && ['query'],
-      datasources: { postgresql: { url: this._url() } },
+      datasources: { [this.provider]: { url: this._url() } },
     });
     await this.prisma.$connect();
   }
@@ -32,7 +33,9 @@ class PrismaAdapter extends BaseKeystoneAdapter {
     // If this isn't what a user wants, they can update `getSchemaName` to return either
     // a different dbSchemaName, or null if they just want to use the DATABASE_URL as it is.
     // TODO: Should we default to 'public' or null?
-    return this.dbSchemaName ? `${this.url}?schema=${this.dbSchemaName}` : this.url;
+    if (this.provider === 'postgresql') {
+      return this.dbSchemaName ? `${this.url}?schema=${this.dbSchemaName}` : this.url;
+    }
   }
 
   async _generateClient(rels) {
@@ -175,9 +178,9 @@ class PrismaAdapter extends BaseKeystoneAdapter {
     );
 
     const header = `
-      datasource postgresql {
+      datasource ${this.provider} {
         url      = env("DATABASE_URL")
-        provider = "postgresql"
+        provider = "${this.provider}"
       }
       generator client {
         provider = "prisma-client-js"
@@ -200,23 +203,25 @@ class PrismaAdapter extends BaseKeystoneAdapter {
   // This will drop all the tables in the backing database. Use wisely.
   async dropDatabase() {
     let migrationNeeded = true;
-    for (const { tablename } of await this.prisma.$queryRaw(
-      `SELECT tablename FROM pg_tables WHERE schemaname='${this.dbSchemaName}'`
-    )) {
-      if (tablename.includes('_Migration')) {
-        migrationNeeded = false;
-      }
-      await this.prisma.$queryRaw(
-        `TRUNCATE TABLE \"${this.dbSchemaName}\".\"${tablename}\" CASCADE;`
-      );
-    }
-    for (const { relname } of await this.prisma.$queryRaw(
-      `SELECT c.relname FROM pg_class AS c JOIN pg_namespace AS n ON c.relnamespace = n.oid WHERE c.relkind='S' AND n.nspname='${this.dbSchemaName}';`
-    )) {
-      if (!relname.includes('_Migration')) {
+    if (this.provider === 'postgresql') {
+      for (const { tablename } of await this.prisma.$queryRaw(
+        `SELECT tablename FROM pg_tables WHERE schemaname='${this.dbSchemaName}'`
+      )) {
+        if (tablename.includes('_Migration')) {
+          migrationNeeded = false;
+        }
         await this.prisma.$queryRaw(
-          `ALTER SEQUENCE \"${this.dbSchemaName}\".\"${relname}\" RESTART WITH 1;`
+          `TRUNCATE TABLE \"${this.dbSchemaName}\".\"${tablename}\" CASCADE;`
         );
+      }
+      for (const { relname } of await this.prisma.$queryRaw(
+        `SELECT c.relname FROM pg_class AS c JOIN pg_namespace AS n ON c.relnamespace = n.oid WHERE c.relkind='S' AND n.nspname='${this.dbSchemaName}';`
+      )) {
+        if (!relname.includes('_Migration')) {
+          await this.prisma.$queryRaw(
+            `ALTER SEQUENCE \"${this.dbSchemaName}\".\"${relname}\" RESTART WITH 1;`
+          );
+        }
       }
     }
